requirements.txt
Django
psycopg2-binary==2.9.9
djangorestframework==3.15.2
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework-gis>=1.0
drf-spectacular==0.27.2
djangorestframework-simplejwt==5.3.1
numpy
scikit-learn
openpyxl